+++ src/engine/Engine.ts
@@ -565,7 +565,7 @@ O|===|* >________________>\n\
       this._logger.debug('Engine viewport is size ' + options.width + ' x ' + options.height);
     } else if (!options.displayMode) {
       this._logger.debug('Engine viewport is fit');
-      displayMode = DisplayMode.Fit;
+      displayMode = DisplayMode.FitScreen;
     }
 
     if (Flags.isEnabled(Experiments.WebGL)) {